Main practice codes+ Lots of optimizations
Added all of the main practice codes, and fixed up the previous code more.

+++ ttyd-tools/rel/include/ttyd/e_atan2.h
@@ -4,7 +4,7 @@ namespace ttyd::e_atan2 {
 	
 extern "C" {
 
-double __ieee754_atan2(float CoordinateX, float CoordinateY);
+double __ieee754_atan2(double CoordinateX, double CoordinateY);
 
 }
